already complete basic lex analysis

+++ compiler/lexical_analyser/utils.h
@@ -1,8 +1,9 @@
 #include <stdlib.h>
 #include <stdio.h>
+#include <memory.h>
 
 enum ERROR_TYPE{FATAL,WARNING,ERROR};
 
 void error(enum ERROR_TYPE type,char *string);
-void *do_malloc(int size);
+void *do_malloc(long size);
 void *do_free(void *p);
